Działa jak wynik ma być floatem

+++ program/include/arithmetic.hpp
@@ -7,7 +7,7 @@ class Arithmetic : public Series {
 public:
     string get_name() override;
     bool is_a(vector<int> given) override;
-    int get_next(int first, int second) override;
+    float get_next(int first, int second) override;
 };
 
 #endif
