+++ program/include/series.hpp
@@ -8,7 +8,7 @@ public:
     virtual ~Series() {};
     virtual std::string get_name() = 0;
     virtual bool is_a(vector<int> given) = 0;
-    virtual int get_next(int first, int second) = 0;
+    virtual float get_next(int first, int second) = 0;
     Prediction *get_prediction(vector<int>);
 };
